Widened CHS_NY note NAs
Hey Bill, you can do the following code:
source("widen_chs_ny.R") and it will run the script and add chs_ny_wide to your environment for working on.

Because there are NAs, it'll be harder to do math on these columns. I did an ugly nested loop to get some correlations, but do some googling and find a formal package for correlation plots.